requirements.txt
fastapi==0.127.1
uvicorn[standard]==0.40.0
python-multipart==0.0.21
pillow==12.0.0
torch>=2.0.0
torchvision>=0.15.0
timm>=1.0.0
